third_party/blink/web_tests/external/wpt/css/cssom/page-descriptors-expected.txt

This is a testharness.js-based test.
[FAIL] Page descriptors shouldn't be exposed to CSS Style declarations
  assert_equals: computed style should not have size property expected (undefined) undefined but got (string) ""
Harness: the test ran to completion.
